requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-migrate>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0; sys_platform != 'win32'
pytest>=7.0.0